Move layerUniforms to shaderlib
