+++ modules/layers/src/bitmap-layer/bitmap-layer-uniforms.ts
@@ -1,6 +1,6 @@
 import type {Texture} from '@luma.gl/core';
 import {ShaderModule} from '@luma.gl/shadertools';
-import {UniformTypes} from '../uniform-types';
+import {UniformTypes} from '@deck.gl/core';
 
 const uniformBlock = `\
 uniform bitmapUniforms {
